+++ actionpack/lib/action_dispatch/http/headers.rb
@@ -1,6 +1,16 @@
 module ActionDispatch
   module Http
     class Headers
+      NON_PREFIX_VARIABLES = %w(
+        CONTENT_TYPE CONTENT_LENGTH
+        HTTPS AUTH_TYPE GATEWAY_INTERFACE
+        PATH_INFO PATH_TRANSLATED QUERY_STRING
+        REMOTE_ADDR REMOTE_HOST REMOTE_IDENT REMOTE_USER
+        REQUEST_METHOD SCRIPT_NAME
+        SERVER_NAME SERVER_PORT SERVER_PROTOCOL SERVER_SOFTWARE
+      )
+      HEADER_REGEXP = /\A[A-Za-z-]+\z/
+
       include Enumerable
 
       def initialize(env = {})
@@ -32,7 +42,9 @@ module ActionDispatch
       end
 
       def cgi_name(k)
-        "HTTP_#{k.upcase.gsub(/-/, '_')}"
+        k = k.upcase.gsub(/-/, '_')
+        k = "HTTP_#{k.upcase.gsub(/-/, '_')}" unless NON_PREFIX_VARIABLES.include?(k)
+        k
       end
     end
   end